Подключает js  модуль к html, производит разбивку js кода на модули
